requirements.txt
django==4.1.6
stripe==5.1.1
gunicorn==20.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
